split Meal component into different files
