+++ src/components/Nav.js
@@ -17,6 +17,7 @@ const Nav = props => (
         <Link to="/welcome">Home</Link>
         <Link to="/">Login</Link>
         <Link to="/signup">Sign Up</Link>
+        <Link to="/meal">Table</Link>
         <Link to="/contacts"> Contacts </Link>
         <Link to='/' onClick={logout}>Sign Out</Link>
       </div>
